refactor: add wrapper dir for components
add ClipCard, ClipList dir for wrapper
