+++ src/Components/ClipList/index.ts
@@ -1,0 +1,3 @@
+import ClipList from "./ClipList"
+
+export default ClipList
